given_gemfile can lock to specific versions.
This cuts down on the "every case already has this pattern" of needing
to restrict the locked versions in play when specifying the Gemfile.
It's got a little too much magic and hopes for simple Gemfiles -- this
may not scale well, but the tool in general is still relatively simple.